+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
-- Model: `gpytorch-qr==0.10.0`
+- Model: `gpytorch-qr==0.11.0`
 - Dataset:
   - `heavyedge/profiles:v1.0.0`
   - `heavyedge/shape-features:v1.1.0`
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hyperparameters are optimized by 5-fold cross validation.
 - GPQR now use GPR posterior mean as prior mean.
 - GPQR now use GPR lengthscale.
+- GPQR predictions now enforce strict quantile ordering without dtype promotion.
 
 ### Removed